split controllers into generalControllers
