+++ server/services/kubernetesService.js
@@ -105,11 +105,6 @@ const kubernetesService = {
             //console.log(error);
             if (error instanceof Error) {
                 return error;
-                // return {
-                //     name: error.name,
-                //     message: error.message,
-                //     //stack: error.stack
-                // } as err;
             }
         }
     },
